+++ src/chain/esplora.rs
@@ -29,6 +29,7 @@ use crate::fee_estimator::{
 };
 use crate::io::utils::write_node_metrics;
 use crate::logger::{log_bytes, log_error, log_info, log_trace, LdkLogger, Logger};
+use crate::tx_broadcaster::TxBroadcastError;
 use crate::types::{ChainMonitor, ChannelManager, DynStore, Sweeper, Wallet};
 use crate::{Error, NodeMetrics};
 
@@ -479,58 +480,73 @@ impl EsploraChainSource {
 		Ok(())
 	}
 
-	pub(crate) async fn process_broadcast_package(&self, package: Vec<Transaction>) {
+	pub(crate) async fn process_broadcast_package(
+		&self, package: Vec<Transaction>,
+	) -> Result<(), TxBroadcastError> {
+		let mut package_result = Ok(());
 		for tx in &package {
 			let txid = tx.compute_txid();
 			let timeout_fut = tokio::time::timeout(
 				Duration::from_secs(TX_BROADCAST_TIMEOUT_SECS),
 				self.esplora_client.broadcast(tx),
 			);
-			match timeout_fut.await {
+			let tx_result = match timeout_fut.await {
 				Ok(res) => match res {
 					Ok(()) => {
 						log_trace!(self.logger, "Successfully broadcast transaction {}", txid);
+						Ok(())
 					},
-					Err(e) => match e {
-						esplora_client::Error::HttpResponse { status, message } => {
-							if status == 400 {
-								// Log 400 at lesser level, as this often just means bitcoind already knows the
-								// transaction.
-								// FIXME: We can further differentiate here based on the error
-								// message which will be available with rust-esplora-client 0.7 and
-								// later.
+					Err(e) => {
+						let result = if matches!(
+							&e,
+							esplora_client::Error::HttpResponse { status: 400, .. }
+						) {
+							Err(TxBroadcastError::Rejected)
+						} else {
+							Err(TxBroadcastError::Failed)
+						};
+						match e {
+							esplora_client::Error::HttpResponse { status, message } => {
+								if status == 400 {
+									// Log 400 at lesser level, as this often just means bitcoind already knows the
+									// transaction.
+									// FIXME: We can further differentiate here based on the error
+									// message which will be available with rust-esplora-client 0.7 and
+									// later.
+									log_trace!(
+										self.logger,
+										"Failed to broadcast due to HTTP connection error: {}",
+										message
+									);
+								} else {
+									log_error!(
+										self.logger,
+										"Failed to broadcast due to HTTP connection error: {} - {}",
+										status,
+										message
+									);
+								}
 								log_trace!(
 									self.logger,
-									"Failed to broadcast due to HTTP connection error: {}",
-									message
+									"Failed broadcast transaction bytes: {}",
+									log_bytes!(tx.encode())
 								);
-							} else {
+							},
+							_ => {
 								log_error!(
 									self.logger,
-									"Failed to broadcast due to HTTP connection error: {} - {}",
-									status,
-									message
+									"Failed to broadcast transaction {}: {}",
+									txid,
+									e
 								);
-							}
-							log_trace!(
-								self.logger,
-								"Failed broadcast transaction bytes: {}",
-								log_bytes!(tx.encode())
-							);
-						},
-						_ => {
-							log_error!(
-								self.logger,
-								"Failed to broadcast transaction {}: {}",
-								txid,
-								e
-							);
-							log_trace!(
-								self.logger,
-								"Failed broadcast transaction bytes: {}",
-								log_bytes!(tx.encode())
-							);
-						},
+								log_trace!(
+									self.logger,
+									"Failed broadcast transaction bytes: {}",
+									log_bytes!(tx.encode())
+								);
+							},
+						}
+						result
 					},
 				},
 				Err(e) => {
@@ -545,9 +561,14 @@ impl EsploraChainSource {
 						"Failed broadcast transaction bytes: {}",
 						log_bytes!(tx.encode())
 					);
+					Err(TxBroadcastError::Timeout)
 				},
+			};
+			if package_result.is_ok() {
+				package_result = tx_result;
 			}
 		}
+		package_result
 	}
 
 	pub(super) async fn get_address_balance(&self, address: &bitcoin::Address) -> Option<u64> {